+++ src/views/DrefApplicationExport/index.tsx
@@ -243,8 +243,8 @@ export function Component() {
         drefResponse?.targeting_strategy_support_file_details,
     );
     const showTargetingStrategySection = peopleAssistedDefined
-    || selectionCriteriaDefined
-    || targetingStrategySupportingDocumentDefined;
+        || selectionCriteriaDefined
+        || targetingStrategySupportingDocumentDefined;
 
     const riskSecurityDefined = isDefined(drefResponse)
         && isDefined(drefResponse.risk_security)
@@ -254,8 +254,8 @@ export function Component() {
         drefResponse?.has_child_safeguarding_risk_analysis_assessment,
     );
     const showRiskAndSecuritySection = riskSecurityDefined
-    || riskSecurityConcernDefined
-    || hasChildrenSafeguardingDefined;
+        || riskSecurityConcernDefined
+        || hasChildrenSafeguardingDefined;
 
     const plannedInterventionDefined = isDefined(drefResponse)
         && isDefined(drefResponse.planned_interventions)
@@ -1007,13 +1007,6 @@ export function Component() {
                             </Heading>
                             <Container>
                                 <TextOutput
-                                    label={strings.drefAllocationLabel}
-                                    value={drefResponse?.amount_requested}
-                                    valueType="number"
-                                    prefix={strings.chfPrefix}
-                                    strongLabel
-                                />
-                                <TextOutput
                                     label={strings.budgetLabel}
                                     value={plannedIntervention.budget}
                                     valueType="number"
@@ -1190,8 +1183,9 @@ export function Component() {
                         )}
                     </Container>
                     <Link
-                        to="emergencies"
+                        href="/emergencies"
                         withUnderline
+                        external
                     >
                         {strings.drefExportReference}
                     </Link>
