+++ src/views/DrefApplicationForm/Operation/i18n.json
@@ -8,6 +8,7 @@
         "drefFormPeopleAssistedThroughOperationDescription": "Explain the logic behind our targets. Which groups are we targeting and why are we targeting these particular groups? Explain how you will target vulnerable groups (e.g., Migrants, refugees, etc.)",
         "drefFormPeopleTargetedWithEarlyActions": "Numbers of persons targeted with early actions (if any)",
         "drefFormPlannedIntervention": "Planned Intervention",
+        "drefFormRequestAmount": "Requested Amount in CHF",
         "drefFormPmer": "How will this operation be monitored?",
         "drefFormPmerDescription": "Will there be IFRC monitoring visits? How will it be deployed?",
         "drefFormResponseRationale": "Operation strategy rationale",
@@ -28,7 +29,7 @@
         "drefFormTargetingStrategy": "Targeting Strategy",
         "drefFormTotal": "Total Population",
         "drefFormWomen": "Women",
-        "drefFormResponseTotalAmountOfPlannedBudget" : "Total amount of planned budget does not match the Requested Amount",
+        "drefFormResponseTotalAmountOfPlannedBudget": "Total amount of planned budget does not match the Requested Amount",
         "drefFormResponseAddButton": "Add",
         "drefFormAssistedPopulation": "Total Targeted Population",
         "drefFormBoys": "Boys (under 18)",
