+++ src/views/DrefFinalReportForm/Overview/i18n.json
@@ -4,7 +4,6 @@
         "finalReportMainDonor": "Main Donors",
         "drefOperationalShareApplicationLabel": "The DREF Final Report is shared with",
         "drefOperationalShareApplicationDescription": "The users will be able to view, edit and add other users.",
-        "finalReportTotalAllocation": "Total DREF Allocation",
         "drefFormSharingHeading": "Sharing",
         "drefFormInstantShareLabel": "Create & Share",
         "drefFormEssentialInformation": "Essential Information",
