+++ src/views/DrefOperationalUpdateExport/index.tsx
@@ -236,8 +236,8 @@ export function Component() {
         drefResponse?.has_child_safeguarding_risk_analysis_assessment,
     );
     const showRiskAndSecuritySection = riskSecurityDefined
-    || riskSecurityConcernDefined
-    || hasChildrenSafeguardingDefined;
+        || riskSecurityConcernDefined
+        || hasChildrenSafeguardingDefined;
 
     const plannedInterventionDefined = isDefined(drefResponse)
         && isDefined(drefResponse.planned_interventions)
@@ -1111,8 +1111,9 @@ export function Component() {
                         )}
                     </Container>
                     <Link
-                        to="emergencies"
+                        href="/emergencies"
                         withUnderline
+                        external
                     >
                         {strings.drefExportReference}
                     </Link>
