+++ src/views/DrefOperationalUpdateForm/Operation/i18n.json
@@ -33,6 +33,9 @@
         "drefFormBoys": "Boys (under 18)",
         "drefFormBudgetTemplateLabel": "Budget template",
         "drefFormBudgetTemplateUploadButtonLabel": "Upload your budget summary",
+        "drefOperationalUpdateAllocationSoFar": "Budget Allocated so far",
+        "drefOperationalUpdateAdditionalAllocationRequested": "Additional Allocation Requested",
+        "drefOperationalUpdateTotalAllocation": "Total DREF Allocation",
         "drefFormCommunication": "Please briefly explain the National Societies communication strategy for this operation.",
         "drefFormCommunicationDescription": "Will the IFRC be supporting with communication? What roles will be involved?",
         "drefFormOperationDescription": "Description",
