+++ src/views/DrefOperationalUpdateForm/Overview/i18n.json
@@ -4,9 +4,6 @@
         "drefOperationalUpdateNumber": "Operational Update Number",
         "drefOperationalShareApplicationLabel": "The DREF Operational Update is shared with",
         "drefOperationalShareApplicationDescription": "The users will be able to view, edit and add other users.",
-        "drefOperationalUpdateAllocationSoFar": "Budget Allocated so far",
-        "drefOperationalUpdateAdditionalAllocationRequested": "Additional Allocation Requested",
-        "drefOperationalUpdateTotalAllocation": "Total DREF Allocation",
         "drefFormSharingHeading": "Sharing",
         "drefFormInstantShareLabel": "Create & Share",
         "drefFormEssentialInformation": "Essential Information",
