+++ src/views/DrefOperationalUpdateForm/Overview/index.tsx
@@ -31,7 +31,6 @@ import { type GoApiResponse } from '#utils/restRequest';
 import {
     stringValueSelector,
 } from '#utils/selectors';
-import { sumSafe } from '#utils/common';
 import useBooleanState from '#hooks/useBooleanState';
 import useGlobalEnums from '#hooks/domain/useGlobalEnums';
 import useCountry from '#hooks/domain/useCountry';
@@ -133,20 +132,6 @@ function Overview(props: Props) {
             setFieldValue(false, 'emergency_appeal_planned');
         }
     }, [setFieldValue]);
-
-    const handleAdditionalAllocationChange = useCallback((
-        additionalAllocation: number | undefined,
-        name: 'additional_allocation',
-    ) => {
-        setFieldValue(additionalAllocation, name);
-        setFieldValue(
-            sumSafe([
-                value.dref_allocated_so_far,
-                additionalAllocation,
-            ]),
-            'total_dref_allocation',
-        );
-    }, [setFieldValue, value.dref_allocated_so_far]);
 
     const handleNSChange = useCallback((nationalSociety: number | undefined) => {
         setFieldValue(nationalSociety, 'national_society');
@@ -417,44 +402,6 @@ function Overview(props: Props) {
                             {strings.drefFormGenerateTitle}
                         </Button>
                     </div>
-                </InputSection>
-                <InputSection
-                    title={strings.drefOperationalUpdateAllocationSoFar}
-                    numPreferredColumns={2}
-                >
-                    <NumberInput
-                        name="dref_allocated_so_far"
-                        value={value.dref_allocated_so_far}
-                        onChange={undefined}
-                        error={error?.dref_allocated_so_far}
-                        disabled={disabled}
-                        readOnly
-                    />
-                </InputSection>
-                <InputSection
-                    title={strings.drefOperationalUpdateAdditionalAllocationRequested}
-                    numPreferredColumns={2}
-                >
-                    <NumberInput
-                        name="additional_allocation"
-                        value={value.additional_allocation}
-                        onChange={handleAdditionalAllocationChange}
-                        error={error?.additional_allocation}
-                        disabled={disabled}
-                    />
-                </InputSection>
-                <InputSection
-                    title={strings.drefOperationalUpdateTotalAllocation}
-                    numPreferredColumns={2}
-                >
-                    <NumberInput
-                        name="total_dref_allocation"
-                        value={value.total_dref_allocation}
-                        onChange={undefined}
-                        error={error?.total_dref_allocation}
-                        disabled={disabled}
-                        readOnly
-                    />
                 </InputSection>
 
                 {value?.type_of_dref !== TYPE_LOAN && (
